Move random functions to test_utils, make messages_util a public module.
